Pcbnew: point editor: move arcs to a behaviour class
This is a good example of an edit behaviour that had a lot
of code in PCB_POINT_EDITOR that was only used for a single
item type. Many of these (now static) functions will probably
find use in other places arcs crop up like filleted shapes
and also in eeschema.

But for now, keep it simpler and keep them near where they
are used.