Pcbnew: shape props dialog: avoid slightly fragile index enum
If this enum is not in order, it will choose the wrong tabs.
We can get the tab index directly from the sizers by going
to the parent, which makes this process automatic.

The CTRL_IDX orders are still possible to de-sync, but at least
that's all in the file and not split between CPP/FBP files.