Pcbnew: include pad snaps for visible layers when not in high-constrast mode
This the 'activeLayers' is a more restricted set of layers than
all visible layers. When high constrast mode is off, snaps should
be available for all visible layers, not just layers in the
'activeLayers' set.

This is similar to the check done in checkVisibility, but as
pads are multi-layer items, they have this check afterwards as
well.

Note - the All Layer/Active Layer Only filtering is handled elsewhere.