Eeschema: always add _unitN to symbol export names
The scheme of adding _N to the filename only when there are multiple
units has two problem:

* It can alias (e.g. a single-unit HDSP-4830_2 and unit 2
  of a multi-unit HDSP-4830) and silently overwrite some units of
  other symbols.
* When processing the files, you have to account for a possibly
  present suffix, which can be hard to tell apart from parts of the
  axctual symbol name.

Change to always adding _unitN, so aliasing is impossible, and
a simple pattern can be used to match the generated filenames.

Demorgan variants can still have a optional suffix, but this
comes after the unit suffix, so the pattern to find that becomes
easier too.

Fixes: https://gitlab.com/kicad/code/kicad/-/issues/18929