Improve initial size of launcher window.
To avoid clipping the launcher buttons, this size was determined by hand
tweaking on Linux so it may be less than optimal on other platforms.  This
only affects the very first launch of a new version.  Once the frame size
is saved in the configuration file, that is the new size until the user
changes it.