Eeschema: symbol SVG export: don't lower-case
This is inconsistent with how footprints work and adds work for
downstream scripts which need to be aware that they have to process the
symbol name in this way.

Symbols named "syma" and "SYMA" can still collide in the same library on
case-insensitive systems, but this should be avoided by warning/errors
at the library level, rather than by silent overwriting on export.

Relates-To: https://gitlab.com/kicad/code/kicad/-/issues/19370