Revert "Pcbnew: reduce snap timeout a fraction"
Now the updating is fixed in e861e48b0d6f02d8671fc837e9d550447e19d1c3,
the hover timeout is shorter because it's not mistakenly waiting a
random-ish time for the next mouse movement. So this adjustment is not
required. In fact, if anything, it may be better to go the other way.

This reverts commit 67a076908637a6ae3b39c5ef03edcb0488b6bc89.