Prevent overplotting copper layers
c4ca7aa161d5590726e3f058f38ed0f020d9bda0 wanted to ensure that drill
marks were visible even when plotting non-copper layers.  However, this
had the effect of changing the plot order depending on whether drill
marks were shown or not.  The current solution is to overplot drill
marks if both copper and non-copper layers are shown.

This remains a sub-optimal solution as buried vias will be overplotted
along with through holes.  It may be a better long-term solution to gate
the plotting of covering pad layers' drill marks to the plot options
that include copper layers on the same page